Refactor: Changed structure of files HeroSection Navbar PricingSection Testimonial Section
